requirements.txt
mcp
pillow>=9.0.0 